Y2020

+++ Migrations/ApplicationDbContextModelSnapshot.cs
@@ -291,8 +291,8 @@ namespace htsApp.Migrations
                         .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-                    b.Property<int?>("Age")
-                        .HasColumnType("int");
+                    b.Property<long?>("Age")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("CTCNumber")
                         .HasColumnType("nvarchar(max)");
@@ -342,11 +342,11 @@ namespace htsApp.Migrations
                     b.Property<string>("NameofSite")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int?>("Nooffemalecondomsissued")
-                        .HasColumnType("int");
+                    b.Property<long?>("Nooffemalecondomsissued")
+                        .HasColumnType("bigint");
 
-                    b.Property<int?>("Noofmalecondomsissued")
-                        .HasColumnType("int");
+                    b.Property<long?>("Noofmalecondomsissued")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("Occupation")
                         .HasColumnType("int");
@@ -395,8 +395,8 @@ namespace htsApp.Migrations
                     b.Property<int>("TypeofCounselling")
                         .HasColumnType("int");
 
-                    b.Property<int?>("Year")
-                        .HasColumnType("int");
+                    b.Property<long?>("Year")
+                        .HasColumnType("bigint");
 
                     b.HasKey("ID");
